feat: Currency UI 연동

+++ Assets/02.Scripts/InGame/Currency/Repositories/ICurrencyRepository.cs
@@ -1,5 +1,5 @@
 public interface ICurrencyRepository
 {
-    void Save(double gold);
-    double Load();
+    void Save(int gold);
+    int Load();
 }
